+++ src/propagate.rs
@@ -10,7 +10,7 @@ use nix::unistd::Pid;
 
 
 
-pub fn propagate_signal_to_all_child_pids(pid: i32, depth: u8, signal_map: SignalMap) -> Result<()> {
+pub fn propagate_signal_to_all_child_pids(pid: u32, depth: u8, signal_map: SignalMap) -> Result<()> {
     let target_pids = get_target_pids(pid, depth)?;
 
     let mut join_handles: Vec<std::thread::JoinHandle<_>> = vec![];
@@ -27,7 +27,7 @@ pub fn propagate_signal_to_all_child_pids(pid: i32, depth: u8, signal_map: Signa
     Ok(())
 }
 
-fn propagate_signal(pid: i32, signal_map: SignalMap) -> Result<()> {
+fn propagate_signal(pid: u32, signal_map: SignalMap) -> Result<()> {
     let listener = SignalListener::new(pid, signal_map.listen_signal);
     listener.listen()?;
 
@@ -35,7 +35,7 @@ fn propagate_signal(pid: i32, signal_map: SignalMap) -> Result<()> {
     for i in 1..=child_pids.len() as u8 {
         for pid in child_pids.get(&(i as usize)).unwrap().iter() {
             let send_signal = signal_map.send_signal();
-            if let Ok(_) = kill(Pid::from_raw(*pid), send_signal) {
+            if let Ok(_) = kill(Pid::from_raw(*pid as i32), send_signal) {
                 println!("Sent {} signal to child process: {} successfully", send_signal, pid);
             }
             else {
@@ -47,7 +47,7 @@ fn propagate_signal(pid: i32, signal_map: SignalMap) -> Result<()> {
     Ok(())
 }
 
-fn get_target_pids(pid: i32, depth: u8) -> Result<ChildPids> {
+fn get_target_pids(pid: u32, depth: u8) -> Result<ChildPids> {
     if depth == 0 {
         return Ok(vec![pid]);
     }
